[node] Add a draft plugin-system implementation - PLEASE REVIEW @aschaeffer
