+++ node/app/plugins.json
@@ -1,0 +1,3 @@
+[
+  "plugins-rest"
+]
