+++ node/app/plugins/rest/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "inexor-plugin-rest",
+  "version": "0.9.0",
+  "description": "A RESTfull frontend for the Inexor tree API",
+  "main": "index.js",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/inexor-game/code.git"
+  },
+  "maintainer": "Lennard Berger <fohlen@darkkeepers.dk>",
+  "license": "Zlib",
+  "bugs": {
+    "url": "https://github.com/inexor-game/code/issues"
+  },
+  "homepage": "https://github.com/inexor-game/inexor/node",
+  "engines": {
+    "node": ">= 4.4.4"
+  }
+}
