--- node/config.json
@@ -1,6 +1,0 @@
-{
-	"name": "Inexor",
-	"version": "0.8.0",
-	"level": "debug",
-	"logFile": "server.log"
-}
